deneme5/ARDA_MARKET/requirements.txt
Pillow
bcrypt
pyotp
qrcode